Level_3/3_12_deploying_agents_with_fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langgraph
requests
langgraph-checkpoint-sqlite